refactor(annotator): extract floating panel shell from find & replace
