+++ packages/client/src/components/advanced/Annotator/AnnotatorSearchLine/AnnotatorFindReplaceModal.tsx
@@ -1,33 +1,16 @@
-import {
-  autoUpdate,
-  flip,
-  FloatingPortal,
-  limitShift,
-  offset,
-  shift,
-  useFloating,
-} from "@floating-ui/react";
 import { Annotator, Occurrence } from "@inkvisitor/annotator/src/lib";
 import { IDocument } from "@inkvisitor/shared/types";
 import { Button, Checkbox, Input, Loader, Submit } from "components";
-import useKeypress from "hooks/useKeyPress";
-import React, { useLayoutEffect, useMemo, useRef, useState } from "react";
-import { FaTimes } from "react-icons/fa";
-import { MdDragIndicator } from "react-icons/md";
+import React, { useMemo, useRef, useState } from "react";
 import { toast } from "react-toastify";
-import { ANNOTATOR_MENU_PAGE_PADDING, useAnnotatorMenuDrag } from "../hooks/useAnnotatorMenuDrag";
+import { AnnotatorFloatingPanel } from "../AnnotatorFloatingPanel/AnnotatorFloatingPanel";
 import { useDocumentContentSave } from "../hooks/useDocumentContentSave";
 import {
-  StyledFindReplaceBody,
   StyledFindReplaceButtonWrap,
-  StyledFindReplaceDraggable,
-  StyledFindReplaceFloating,
   StyledFindReplaceFlags,
   StyledFindReplaceFooter,
-  StyledFindReplaceHeader,
   StyledFindReplaceResults,
   StyledFindReplaceRow,
-  StyledFindReplaceTitle,
   StyledNoResults,
 } from "./AnnotatorFindReplaceModalStyles";
 import { applyReplacements, nextActiveOccurenceIndex, ReplaceRange } from "./replaceUtils";
@@ -91,60 +74,10 @@ export const AnnotatorFindReplaceModal: React.FC<AnnotatorFindReplaceModal> = ({
   const [isReplacingAll, setIsReplacingAll] = useState<boolean>(false);
   const [showReplaceAllSubmit, setShowReplaceAllSubmit] = useState<boolean>(false);
 
-  const { dragHandleProps, draggableRef, dragOffset } = useAnnotatorMenuDrag();
-
-  // Opens centered over the page, like the annotator selection menu — the user
-  // drags it out of the way from there.
-  const middleware = useMemo(() => {
-    if (typeof document === "undefined") return [];
-    const page = document.getElementById("page");
-    const centerOnPoint = offset(({ rects }) => ({
-      mainAxis: -(rects.floating.height || 0) / 2,
-    }));
-    if (!page) return [centerOnPoint];
-    return [
-      centerOnPoint,
-      flip({
-        boundary: page,
-        padding: ANNOTATOR_MENU_PAGE_PADDING,
-      }),
-      shift({
-        boundary: page,
-        padding: ANNOTATOR_MENU_PAGE_PADDING,
-        crossAxis: true,
-        limiter: limitShift(),
-      }),
-    ];
-  }, []);
-
-  const { refs, floatingStyles } = useFloating({
-    open: true,
-    placement: "bottom",
-    strategy: "fixed",
-    whileElementsMounted: autoUpdate,
-    middleware,
-  });
-
-  useLayoutEffect(() => {
-    const page = document.getElementById("page");
-    if (!page) return;
-    refs.setPositionReference({
-      getBoundingClientRect() {
-        const r = page.getBoundingClientRect();
-        const cx = r.left + r.width / 2;
-        const cy = r.top + r.height / 2;
-        return new DOMRect(cx, cy, 0, 0);
-      },
-      contextElement: page,
-    });
-  }, []);
-
   // Focus lands on whichever field the user still has to fill in. Evaluated as
   // the inputs mount, which FloatingPortal defers to its second render pass —
   // the two flags are mutually exclusive, so exactly one input claims focus.
   const focusFindInput = searchTerm.length === 0;
-
-  useKeypress("Escape", onClose);
 
   const saveDocumentContent = useDocumentContentSave({
     annotator,
@@ -237,149 +170,118 @@ export const AnnotatorFindReplaceModal: React.FC<AnnotatorFindReplaceModal> = ({
 
   return (
     <>
-      <FloatingPortal id="page">
-        <StyledFindReplaceFloating
-          ref={(node) => {
-            refs.setFloating(node);
-          }}
-          style={floatingStyles}
-        >
-          <StyledFindReplaceDraggable
-            ref={draggableRef}
-            style={{
-              transform: `translate(${dragOffset.x}px, ${dragOffset.y}px)`,
+      <AnnotatorFloatingPanel
+        title="Find & replace"
+        onClose={onClose}
+        closeTooltipLabel="close find & replace (Esc)"
+      >
+        <StyledFindReplaceRow>
+          <Input
+            value={searchTerm}
+            onChangeFn={(newText: string) => setSearchTerm(newText)}
+            onEnterPressFn={goToNextOccurence}
+            onEscapePressFn={onClose}
+            changeOnType
+            clearable
+            autoFocus={focusFindInput}
+            width="full"
+            inputRef={findInputRef}
+            placeholder="Find"
+            rightContent={
+              <StyledFindReplaceResults>
+                {searchOccurences === null ? (
+                  ""
+                ) : occurencesCount === 0 ? (
+                  <StyledNoResults>no results</StyledNoResults>
+                ) : (
+                  `${searchActiveOccurence + 1} of ${occurencesCount}`
+                )}
+              </StyledFindReplaceResults>
+            }
+          />
+        </StyledFindReplaceRow>
+
+        <StyledFindReplaceRow>
+          <Input
+            value={replaceWith}
+            onChangeFn={(value: string) => setReplaceWith(value)}
+            onEnterPressFn={() => {
+              if (!actionsDisabled) {
+                replaceOccurence();
+              }
             }}
-          >
-            <StyledFindReplaceHeader>
-              {/* Drag lives on the title only — the handle's onPointerDown
-                  preventDefault()s, which would swallow clicks on the close
-                  button if it sat inside the draggable area. */}
-              <StyledFindReplaceTitle {...dragHandleProps}>
-                <MdDragIndicator size={16} />
-                Find & replace
-              </StyledFindReplaceTitle>
-              <Button
-                icon={<FaTimes size={12} />}
-                color="primary"
-                inverted
-                noBorder
-                noBackground
-                onClick={onClose}
-                tooltipLabel="close find & replace (Esc)"
-              />
-            </StyledFindReplaceHeader>
+            onEscapePressFn={onClose}
+            changeOnType
+            clearable
+            autoFocus={!focusFindInput}
+            width="full"
+            inputRef={replaceInputRef}
+            placeholder="Replace with"
+          />
+        </StyledFindReplaceRow>
 
-            <StyledFindReplaceBody>
-              <StyledFindReplaceRow>
-                <Input
-                  value={searchTerm}
-                  onChangeFn={(newText: string) => setSearchTerm(newText)}
-                  onEnterPressFn={goToNextOccurence}
-                  onEscapePressFn={onClose}
-                  changeOnType
-                  clearable
-                  autoFocus={focusFindInput}
-                  width="full"
-                  inputRef={findInputRef}
-                  placeholder="Find"
-                  rightContent={
-                    <StyledFindReplaceResults>
-                      {searchOccurences === null ? (
-                        ""
-                      ) : occurencesCount === 0 ? (
-                        <StyledNoResults>no results</StyledNoResults>
-                      ) : (
-                        `${searchActiveOccurence + 1} of ${occurencesCount}`
-                      )}
-                    </StyledFindReplaceResults>
-                  }
-                />
-              </StyledFindReplaceRow>
+        <StyledFindReplaceFlags>
+          <Checkbox
+            label="Match case"
+            value={isCaseSensitiveMode}
+            onChangeFn={(checked: boolean) => setIsCaseSensitiveMode(checked)}
+            size={13}
+          />
+          <Checkbox
+            label="Whole word only"
+            value={isWholeWordOnlyMode}
+            onChangeFn={(checked: boolean) => setIsWholeWordOnlyMode(checked)}
+            size={13}
+          />
+          <Checkbox
+            label="Use regular expressions"
+            value={isRegexMode}
+            onChangeFn={(checked: boolean) => setIsRegexMode(checked)}
+            size={13}
+          />
+        </StyledFindReplaceFlags>
 
-              <StyledFindReplaceRow>
-                <Input
-                  value={replaceWith}
-                  onChangeFn={(value: string) => setReplaceWith(value)}
-                  onEnterPressFn={() => {
-                    if (!actionsDisabled) {
-                      replaceOccurence();
-                    }
-                  }}
-                  onEscapePressFn={onClose}
-                  changeOnType
-                  clearable
-                  autoFocus={!focusFindInput}
-                  width="full"
-                  inputRef={replaceInputRef}
-                  placeholder="Replace with"
-                />
-              </StyledFindReplaceRow>
-
-              <StyledFindReplaceFlags>
-                <Checkbox
-                  label="Match case"
-                  value={isCaseSensitiveMode}
-                  onChangeFn={(checked: boolean) => setIsCaseSensitiveMode(checked)}
-                  size={13}
-                />
-                <Checkbox
-                  label="Whole word only"
-                  value={isWholeWordOnlyMode}
-                  onChangeFn={(checked: boolean) => setIsWholeWordOnlyMode(checked)}
-                  size={13}
-                />
-                <Checkbox
-                  label="Use regular expressions"
-                  value={isRegexMode}
-                  onChangeFn={(checked: boolean) => setIsRegexMode(checked)}
-                  size={13}
-                />
-              </StyledFindReplaceFlags>
-
-              <StyledFindReplaceFooter>
-                <Button
-                  label="Previous"
-                  color="info"
-                  inverted
-                  tooltipLabel="previous occurence"
-                  tooltipContent={<p>(Shift + F3)</p>}
-                  onClick={goToPreviousOccurence}
-                  disabled={occurencesCount === 0}
-                />
-                <Button
-                  label="Next"
-                  color="info"
-                  inverted
-                  tooltipLabel="next occurence"
-                  tooltipContent={<p>(F3)</p>}
-                  onClick={goToNextOccurence}
-                  disabled={occurencesCount === 0}
-                />
-                <StyledFindReplaceButtonWrap>
-                  <Button
-                    label="Replace"
-                    color="info"
-                    tooltipLabel="replace the current occurence"
-                    onClick={replaceOccurence}
-                    disabled={actionsDisabled}
-                  />
-                  <Loader show={isReplacingOne} size={12} noBackground />
-                </StyledFindReplaceButtonWrap>
-                <StyledFindReplaceButtonWrap>
-                  <Button
-                    label="Replace all"
-                    color="danger"
-                    tooltipLabel="replace every occurence in the document"
-                    onClick={() => setShowReplaceAllSubmit(true)}
-                    disabled={actionsDisabled}
-                  />
-                  <Loader show={isReplacingAll} size={12} noBackground />
-                </StyledFindReplaceButtonWrap>
-              </StyledFindReplaceFooter>
-            </StyledFindReplaceBody>
-          </StyledFindReplaceDraggable>
-        </StyledFindReplaceFloating>
-      </FloatingPortal>
+        <StyledFindReplaceFooter>
+          <Button
+            label="Previous"
+            color="info"
+            inverted
+            tooltipLabel="previous occurence"
+            tooltipContent={<p>(Shift + F3)</p>}
+            onClick={goToPreviousOccurence}
+            disabled={occurencesCount === 0}
+          />
+          <Button
+            label="Next"
+            color="info"
+            inverted
+            tooltipLabel="next occurence"
+            tooltipContent={<p>(F3)</p>}
+            onClick={goToNextOccurence}
+            disabled={occurencesCount === 0}
+          />
+          <StyledFindReplaceButtonWrap>
+            <Button
+              label="Replace"
+              color="info"
+              tooltipLabel="replace the current occurence"
+              onClick={replaceOccurence}
+              disabled={actionsDisabled}
+            />
+            <Loader show={isReplacingOne} size={12} noBackground />
+          </StyledFindReplaceButtonWrap>
+          <StyledFindReplaceButtonWrap>
+            <Button
+              label="Replace all"
+              color="danger"
+              tooltipLabel="replace every occurence in the document"
+              onClick={() => setShowReplaceAllSubmit(true)}
+              disabled={actionsDisabled}
+            />
+            <Loader show={isReplacingAll} size={12} noBackground />
+          </StyledFindReplaceButtonWrap>
+        </StyledFindReplaceFooter>
+      </AnnotatorFloatingPanel>
 
       <Submit
         title="Replace all"
